chore: split off long-running warp-deploy test (#5293)
### Description

chore: split off long-running warp-deploy test

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<table>
<tr>
<th>Before: 14m</th>
<th>After: 7m</th>
</tr>
<tr>
<td>
<img
src="https://github.com/user-attachments/assets/f3b305e4-262a-4f8f-973a-f26790884b18"
alt="Before">
</td>
<td>
<img
src="https://github.com/user-attachments/assets/914d1acd-dcdd-47b5-b338-4ff26c8b0c5a"
alt="After">
</td>
</tr>
</table>
