+++ .github/workflows/test.yml
@@ -136,8 +136,10 @@ jobs:
           - relay
           # Warp Commands
           - warp-apply
+          - warp-apply-extend
           - warp-check
-          - warp-deploy
+          - warp-bridge-1
+          - warp-bridge-2
           - warp-init
           - warp-read
           - warp-send
